+++ Changelog.md
@@ -9,8 +9,10 @@
 * Ability to get file list in project
 * Support of `rubocop-performance`
 * Update rubocop configs to `0.84.0`
+* Add `WrataApi#add_to_queue` default value for `browser`
 
 ### Fixes
+
 * Raise error (not show strange error) while trying to create a lot of server
 
 ### Changes